update-package-with-nodemailer-and-test-email
