+++ api/package.json
@@ -21,6 +21,7 @@
     "mercadopago": "^1.5.17",
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
+    "nodemailer": "^6.9.4",
     "nodemon": "^3.0.1",
     "pg": "^8.11.2",
     "pg-hstore": "^2.3.4",
